services/f2_simulator/requirements.txt
paho-mqtt==1.6.1
python-json-logger==2.0.7